requirements.txt
moviepy==1.0.3
openai==1.3.8
whisper-timestamped==1.14.2
requests==2.31.0
python-dotenv==1.0.0
flask>=2.0.0
werkzeug>=2.0.0
